wellness_wave_backend/requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
python-dotenv
pandas
numpy
scikit-learn
joblib
